Add LSP and other sensible things
Tried to use pyright but that was not working properly.
Switched to using basedpyright and it's working great.
Added autocommand for deactivating automatic comment when going into the
next line.
Add map to keep in visual mode when identing.
Used gitsigns defaults, which are much prettier than whathever was
suggested.
Set navigation bar.
Set transparency.